Update Fast Gradient Sign Method (FSGM)
This Fast Gradient Sign Method (FSGM) algorithm attempts to fool the classifier/model detailed in the main algorithm by applying a range of many minute changes in the input image that changes the classifier's determination all while being imperceptible to the human eye. 

First, this is done by changing the input image data to that of an array so that it can be properly modified in future lines of code. Then, said changes to the input image are modified through a gradient tape which is then used to calculate gradient, sign of gradient, and loss. After this, the modified "adversarial" image is produced using the new signed gradient as well as a small scalar value called epsilon that decides the magnitude of the changes to the input image. 

With the adversarial image created, now it must be changed back to an acceptable pixel range before being converted to the original pixel value range that can be processed by the model.

Finally, this new algorithm is tested on the main algorithm by loading the model and then the images from the folder. Then, the small constant scalar value must be defined. Finally, the FGSM command/attack is done to create the adversarial image which is then saved.